+++ kartverketprosjekt/Services/Admin/AdminService.cs
@@ -118,17 +118,6 @@ namespace kartverketprosjekt.Services.Admin
             return (true, $"{user.epost} ble slettet.");
         }
     }
-
-    public class AdminStats
-    {
-        public int CaseCount { get; set; }
-        public int UserCount { get; set; }
-        public int OpenCasesUnbehandlet { get; set; }
-        public int OpenCasesUnderBehandling { get; set; }
-        public int OpenCasesAvvist { get; set; }
-        public int OpenCasesArkivert { get; set; }
-        public int ClosedCases { get; set; }
-        public List<BrukerModel> Users { get; set; }
-    }
+    
 }
